pyproject.toml
[tool.poetry]
name = "djangorave"
version = "0.1.2"
description = "Django integration for Flutterwave Rave Card payments and subscriptions"
authors = ["Brendon Delate <brendon.delate@gmail.com>"]
license = "MIT"
repository = "https://github.com/bdelate/django-rave.git"
homepage = "https://github.com/bdelate/django-rave.git"
readme = "README.md"

keywords=[
    "flutterwave",
    "rave",
	"django",
]

classifiers=[
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries",
    "Topic :: Utilities",
    "Environment :: Web Environment",
    "Framework :: Django",
]

packages = [
    { include = "djangorave" },
]

[tool.poetry.dependencies]
python = ">=3.6 <4"
django = ">=2.2 <4"
requests = "^2.22"
djangorestframework = "^3.10"

[tool.poetry.dev-dependencies]
black = {version = "=19.3b0",allow-prereleases = true}
mypy = "^0.720.0"
twine = "^1.13"
coverage = "^4.5"
django-extensions = "^2.2"
jupyter = "^1.0"
factory-boy = "^2.12"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
